requirements.txt
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
pdfminer.six==20260107
pycparser==3.0
sqlparse==0.5.5
